# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AstroShield API",
    "description": "The AstroShield API provides access to satellite data, collision avoidance maneuvers, and space domain awareness analytics.",
    "version": "1.0.0",
    "contact": {
      "email": "support@astroshield.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.astroshield.com/api/v1",
      "description": "Production server"
    },
    {
      "url": "https://api-staging.astroshield.com/api/v1",
      "description": "Staging server"
    },
    {
      "url": "http://localhost:3001/api/v1",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/maneuvers": {
      "get": {
        "tags": ["maneuvers"],
        "summary": "List maneuvers",
        "description": "Get a list of maneuvers, optionally filtered by satellite ID and status",
        "operationId": "listManeuvers",
        "parameters": [
          {
            "name": "satellite_id",
            "in": "query",
            "description": "Filter by satellite ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by maneuver status",
            "schema": {
              "type": "string",
              "enum": ["SCHEDULED", "IN_PROGRESS", "COMPLETED", "FAILED", "CANCELLED"]
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Items per page",
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of maneuvers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Maneuver"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PaginationMeta"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["maneuvers"],
        "summary": "Create maneuver",
        "description": "Create a new maneuver for a satellite",
        "operationId": "createManeuver",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManeuverCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Maneuver created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Maneuver"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Satellite not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ManeuverType": {
        "type": "string",
        "enum": ["COLLISION_AVOIDANCE", "STATION_KEEPING", "ORBIT_RAISING", "ORBIT_LOWERING", "DEORBIT", "INCLINATION_CHANGE", "PHASING"],
        "description": "Type of satellite maneuver"
      },
      "ManeuverStatus": {
        "type": "string",
        "enum": ["SCHEDULED", "IN_PROGRESS", "COMPLETED", "FAILED", "CANCELLED"],
        "description": "Status of a maneuver"
      },
...
      "DirectionVector": {
        "type": "object",
        "required": ["x", "y", "z"],
        "properties": {
          "x": {
            "type": "number",
            "format": "float",
            "description": "X component of direction vector",
            "example": 0.1
          },
          "y": {
            "type": "number",
            "format": "float",
            "description": "Y component of direction vector",
            "example": 0.0
          },
          "z": {
            "type": "number",
            "format": "float",
            "description": "Z component of direction vector",
            "example": -0.1
          }
        }
      },
      "TargetOrbit": {
        "type": "object",
        "properties": {
          "altitude": {
            "type": "number",
            "format": "float",
            "description": "Target altitude in kilometers",
            "example": 500.2
          },
          "inclination": {
            "type": "number",
            "format": "float",
            "description": "Target inclination in degrees",
            "example": 45.0
          },
          "eccentricity": {
            "type": "number",
            "format": "float",
            "description": "Target eccentricity",
            "example": 0.001
          }
        }
      },
      "ManeuverParameters": {
        "type": "object",
        "required": ["delta_v", "burn_duration", "direction"],
        "properties": {
          "delta_v": {
            "type": "number",
            "format": "float",
            "description": "Change in velocity (km/s)",
            "example": 0.02
          },
          "burn_duration": {
            "type": "number",
            "format": "float",
            "description": "Duration of thruster burn (seconds)",
            "example": 15.0
          },
          "direction": {
            "$ref": "#/components/schemas/DirectionVector"
          },
          "target_orbit": {
            "$ref": "#/components/schemas/TargetOrbit"
          }
        }
      },
      "ManeuverResources": {
        "type": "object",
        "properties": {
          "fuel_remaining": {
            "type": "number",
            "format": "float",
            "description": "Fuel remaining after maneuver (kg)",
            "example": 24.5
          },
          "power_available": {
            "type": "number",
            "format": "float",
            "description": "Power available during maneuver (W)",
            "example": 1800.0
          },
          "thruster_status": {
            "type": "string",
            "description": "Status of thrusters",
            "example": "NOMINAL"
          }
        }
      },
...
      "ManeuverCreate": {
        "type": "object",
        "required": ["satellite_id", "type", "scheduled_start_time", "parameters"],
        "properties": {
          "satellite_id": {
            "type": "string",
            "description": "ID of the satellite performing the maneuver",
            "example": "sat-001"
          },
          "type": {
            "$ref": "#/components/schemas/ManeuverType"
          },
          "scheduled_start_time": {
            "type": "string",
            "format": "date-time",
            "description": "Planned start time (ISO 8601)",
            "example": "2023-06-15T20:00:00Z"
          },
          "parameters": {
            "$ref": "#/components/schemas/ManeuverParameters"
          }
        }
      },
      "Maneuver": {
        "type": "object",
        "required": [
          "id",
          "satellite_id",
          "type",
          "status",
          "scheduled_start_time",
          "parameters",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the maneuver",
            "example": "mnv-001"
          },
          "satellite_id": {
            "type": "string",
            "description": "ID of the satellite performing the maneuver",
            "example": "sat-001"
          },
          "type": {
            "$ref": "#/components/schemas/ManeuverType"
          },
          "status": {
            "$ref": "#/components/schemas/ManeuverStatus"
          },
          "scheduled_start_time": {
            "type": "string",
            "format": "date-time",
            "description": "Planned start time (ISO 8601)",
            "example": "2023-06-15T20:00:00Z"
          },
          "actual_start_time": {
            "type": "string",
            "format": "date-time",
            "description": "Actual start time if executed",
            "example": "2023-06-15T20:00:15Z"
          },
          "end_time": {
            "type": "string",
            "format": "date-time",
            "description": "End time if completed",
            "example": "2023-06-15T20:00:30Z"
          },
          "parameters": {
            "$ref": "#/components/schemas/ManeuverParameters"
          },
          "resources": {
            "$ref": "#/components/schemas/ManeuverResources"
          },
          "created_by": {
            "type": "string",
            "description": "User who created the maneuver",
            "example": "operator@astroshield.com"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the record was created",
            "example": "2023-06-10T15:30:00Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the record was last updated",
            "example": "2023-06-10T15:30:00Z"
          }
        }
      },
      "PaginationMeta": {
        "type": "object",
        "required": ["page", "limit", "total", "pages"],
        "properties": {
          "page": {
            "type": "integer",
            "description": "Current page number",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "description": "Items per page",
            "example": 20
          },
          "total": {
            "type": "integer",
            "description": "Total number of items",
            "example": 45
          },
          "pages": {
            "type": "integer",
            "description": "Total number of pages",
            "example": 3
          }
        }
      },
      "Error": {
        "type": "object",
        "required": ["detail", "status_code"],
        "properties": {
          "detail": {
            "type": "string",
            "description": "Human-readable error message",
            "example": "Satellite with ID 'sat-999' not found"
          },
          "status_code": {
            "type": "integer",
            "description": "HTTP status code",
            "example": 404
          },
          "error_code": {
            "type": "string",
            "description": "Machine-readable error code",
            "example": "resource_not_found"
          },
          "path": {
            "type": "string",
            "description": "Path that caused the error",
            "example": "/api/v1/satellites/sat-999"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Time when the error occurred",
            "example": "2023-06-18T14:25:30Z"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from the /token endpoint"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key provided by AstroShield"
      }
    }
  }
} 